FIxed code
The fix involves adjusting the sensitivity of the mouse scroll wheel input in the `SwitchGun` method of the `playerController` script. This is achieved by introducing a `scrollSensitivity` factor that scales the raw scroll input. By adjusting this factor, developers can control how much the scroll wheel input influences the selection of guns in the game.