Added Leveling System
Added a leveling system to check if the player has gained enough experience points to level up. If the player has it will level the player one level automatically at this time. We can add/ change that if we want later. I also added where the player obtains experience points for picking up items, weapons and kills an enemy.